% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp-latte.R
\docType{data}
\name{ctp_latte}
\alias{ctp_latte}
\title{Catppuccin latte colors}
\format{
An object of class \code{character} of length 26.
}
\usage{
ctp_latte
}
\description{
Catppuccin latte colors
}
\keyword{datasets}
