% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp_macchiato.R
\docType{data}
\name{ctp_macchiato}
\alias{ctp_macchiato}
\title{Catppuccin macchiato colors}
\format{
An object of class \code{character} of length 26.
}
\usage{
ctp_macchiato
}
\description{
Catppuccin macchiato colors
}
\keyword{datasets}
