% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp-mocha.R
\docType{data}
\name{ctp_mocha}
\alias{ctp_mocha}
\title{Catppuccin mocha colors}
\format{
An object of class \code{character} of length 26.
}
\usage{
ctp_mocha
}
\description{
Catppuccin mocha colors
}
\keyword{datasets}
