% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-theme-ctp.R
\name{gt_theme_ctp}
\alias{gt_theme_ctp}
\title{catpuccin gt theme}
\usage{
gt_theme_ctp(gt_table, ctp_theme)
}
\arguments{
\item{gt_table}{a gt object}

\item{ctp_theme}{a catppuccin ctp theme}
}
\description{
catpuccin gt theme
}
