% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_ctp.R
\name{scale_color_ctp}
\alias{scale_color_ctp}
\title{Add Catppuccin colors to your ggplot}
\usage{
scale_color_ctp(ctp_theme)
}
\arguments{
\item{ctp_theme}{a catppuccin ctp theme}
}
\description{
Add Catppuccin colors to your ggplot
}
