% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_ctp.R
\name{scale_fill_ctp}
\alias{scale_fill_ctp}
\title{Fill items using a catppuccin theme.}
\usage{
scale_fill_ctp(ctp_theme)
}
\arguments{
\item{ctp_theme}{a catppuccin ctp theme}
}
\description{
Fill items using a catppuccin theme.
}
