% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ctp_theme.R
\name{show_ctp_theme}
\alias{show_ctp_theme}
\title{Display a catpuccin theme}
\usage{
show_ctp_theme(ctp_theme)
}
\arguments{
\item{ctp_theme}{a catppuccin ctp theme}
}
\description{
Display a catpuccin theme
}
