% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ctp_grid.R
\name{theme_ctp_grid}
\alias{theme_ctp_grid}
\title{Catppuccin ggplot2 theme - with grids}
\usage{
theme_ctp_grid(ctp_theme)
}
\arguments{
\item{ctp_theme}{a catppuccin ctp theme}
}
\description{
Catppuccin ggplot2 theme - with grids
}
