% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp_frappe.R
\docType{data}
\name{ctp_frappe}
\alias{ctp_frappe}
\title{Catppuccin frappe colors}
\format{
An object of class \code{character} of length 26.
}
\usage{
ctp_frappe
}
\description{
Catppuccin frappe colors
}
\keyword{datasets}
